moved the ocr, filtering and validation steps to a new class
