+++ src/services/index.ts
@@ -1,7 +1,6 @@
 export { ErrorBoundary } from './errorBoundary/ErrorBoundary';
 export { NotificationHandler } from './notification/NotificationListener';
 export { isBaseApiError, BaseApiClientError } from './api/base/error';
-export { ocrRead } from './api/ocr/computerVision';
 export {
   useTagScanStatus,
   TagScanningStages
